refactor navigation to use objects
