+++ conveyor/src/main/kotlin/com/hodamohammadi/conveyor/fragments/ChatsListFragment.kt
@@ -13,7 +13,8 @@ import com.hodamohammadi.conveyor.utils.AppUtils
 import com.hodamohammadi.conveyor.viewmodels.ChatViewModel
 import com.hodamohammadi.conveyor.viewmodels.ViewModelFactory
 import com.hodamohammadi.navigation.RoutePath
-import com.hodamohammadi.navigation.loadIntentOrNull
+import com.hodamohammadi.navigations.features.ChatNavigation
+import com.hodamohammadi.navigations.loaders.loadIntentOrNull
 import com.hodamohammadi.services.BaseResourceObserver
 import com.stfalcon.chatkit.commons.models.IDialog
 import com.stfalcon.chatkit.commons.models.IMessage
@@ -80,7 +81,7 @@ class ChatsListFragment : Fragment(), DialogsListAdapter.OnDialogClickListener<I
         // TODO: Implement chat delete/edits
     }
 
-    private fun startChat() = RoutePath.CHAT_ACTIVITY.loadIntentOrNull().let {
+    private fun startChat() = ChatNavigation.dynamicStart?.let {
         it!!.action = RoutePath.SINGLE_CHAT_FRAGMENT
         startActivity(it)
     }
