+++ navigations/src/main/kotlin/com/hodamohammadi/navigations/features/DynamicFeature.kt
@@ -1,0 +1,5 @@
+package com.hodamohammadi.navigations.features
+
+interface DynamicFeature<T> {
+    val dynamicStart: T?
+}
