+++ navigations/src/main/kotlin/com/hodamohammadi/navigations/loaders/IntentLoader.kt
@@ -1,4 +1,4 @@
-package com.hodamohammadi.navigation
+package com.hodamohammadi.navigations.loaders
 
 import android.content.Intent
